refactored MarkerPopup
