+++ src/client/lib/Enums.js
@@ -1,0 +1,9 @@
+module.exports = {
+    visitationStatuses: {
+    	0: '<Pick status>',
+	    reached: 'Reached',
+	    unreachable: 'Existing, but unreachable',
+	    unsure: 'Unreachable, unknown',
+	    missing: 'Confirmed, missing'
+	}
+};
